+++ SCI/src/library_fixed_uniform.h
@@ -155,7 +155,7 @@ void ScaleUp(int32_t size, intType *arr, int32_t sf);
 
 void ConnectAndSetUp(bool use_heliks = false);
 
-void GenerateTriples(int buffer_size = 0, int chunk_size = 0);
+void GenerateTriples(uint64_t buffer_size = 0, uint64_t chunk_size = 0);
 
 void StartComputation(bool use_heliks = false, bool use_low_round = false);
